add validation permissions

+++ app/crud/users.py
@@ -29,9 +29,30 @@ def handle_database_user_error(handler):
     return wrapper
 
 
+def validate_user_permissions(db, caller_id, user_id=None):
+    if (user_id == caller_id and user_id is not None):
+        return
+
+    db_caller = get_user(db, caller_id)
+    if db_caller.is_superuser:
+        return
+    else:
+        raise HTTPException(
+            status_code=403,
+            detail="Not permission for this method"
+        )
+
+
 @handle_database_user_error
 def get_user(db: Session, user_id: int):
     return db.query(UserModel).filter(UserModel.id == user_id).one()
+
+
+@handle_database_user_error
+def get_user_by_id(db: Session, user_id: str, caller_id: str):
+    validate_user_permissions(db, caller_id, user_id=user_id)
+
+    return get_user(db, user_id)
 
 
 @handle_database_user_error
@@ -64,7 +85,9 @@ def update_user(db: Session, user_updated: UserSchemaWithId):
 
 
 @handle_database_user_error
-def delete_user(db: Session, user_id: str):
+def delete_user(db: Session, user_id: str, caller_id: str):
+    validate_user_permissions(db, caller_id, user_id=user_id)
+
     # check if user exists
     user = get_user(db, user_id)
     db.delete(user)
